+++ server-data/resources/[esx_addons]/esx_license/fxmanifest.lua
@@ -3,8 +3,8 @@ fx_version 'adamant'
 game 'gta5'
 
 description 'ESX License'
-
-version '0.0.4'
+lua54 'yes'
+version '1.0.0'
 
 server_scripts {
 	'@es_extended/imports.lua',