+++ .github/CHANGELOG.md
@@ -13,3 +13,9 @@
    [esx_property\locales\en.lua & it.lua]: corrected typos @bitpredator
    [esx_property\server\main.lua]: correction of lint errors @bitpredator
    [esx_property\config.lua]: correction of lint errors @bitpredator
+3. [esx_license\server\main.lua]: Replaced print in esx_license:addLicense
+   [esx_license\fxmanifest.lua]: Update fxmanifest.lua to version 1.0.0
+   [esx_license\README.md]: update reference links in README.md
+   [esx_license\LICENSE]: updates the copyright reference year
+
+